+++ Vehicles/src/com/endava/models/vehicles/contracts/Vehicle.java
@@ -1,13 +1,12 @@
 package com.endava.models.vehicles.contracts;
 
 import com.endava.models.vehicles.enums.VehicleType;
-import java.math.BigDecimal;
 
 public interface Vehicle {
 
   int getLoadCapacity();
 
-  BigDecimal getPricePerKilometer();
+  double getPricePerKilometer();
 
   VehicleType getType();
